Initial commit of ZeroMQ service demo plugin
Initial commit of ZeroMQ service client plugin demo.

The basic idea is to have a service exposed through a ZeroMQ interface,
responding on a `zmq.REP` socket to the following two messages:

 - `start` -> `started`
 - `notify_completion` -> `completed`

This provides a basic inter-process method for starting a service,
followed by a notification upon completion.  The plugin provides a
timeout option for each step in the DMF protocol to limit the amount of
time spent waiting for a service to complete.

__NB__ As of this commit, the MicroDrop app seems to freeze in the case
where the service times out, after the error message indicating the
protocol has failed is displayed.  This is likely due to threading
issues, which will be resolved after the plug-ins are ported to run in
separate processes.